requirements.txt
streamlit
torch
torchvision
Pillow
opencv-python
notion-client